facts update n error fix
