+++ test/integration/targets/ios_interfaces/tests/cli/_populate_config.yaml
@@ -8,7 +8,10 @@
       description this is interface1
       mtu 65
       speed 10
+      no shutdown
       interface GigabitEthernet 0/2
       description this is interface2
+      mtu 110
       duplex auto
       speed 100
+      shutdown
