+++ test/integration/targets/ios_interfaces/tests/cli/merged.yaml
@@ -4,23 +4,6 @@
 
 - include_tasks: _remove_config.yaml
 
-- set_fact:
-    expected_output:
-      - duplex: auto
-        name: GigabitEthernet0/0
-        speed: auto
-      - description: Configured and Merged by Ansible Network
-        duplex: full
-        mtu: 110
-        name: GigabitEthernet0/1
-        speed: 10
-      - description: Configured and Merged by Ansible Network
-        duplex: full
-        mtu: 2800
-        enabled: false
-        name: GigabitEthernet0/2
-        speed: 100
-
 - block:
     - name: Merge provided configuration with device configuration
       ios_interfaces: &merged
@@ -29,7 +12,8 @@
             description: 'Configured and Merged by Ansible Network'
             mtu: 110
             enabled: True
-          - name: GigabitEthernet0/1
+            duplex: half
+          - name: GigabitEthernet0/2
             description: 'Configured and Merged by Ansible Network'
             mtu: 2800
             enabled: False
@@ -39,12 +23,26 @@
       register: result
 
     - debug:
-        msg: "{{ expected_output | symmetric_difference(result['after']) }}"
+        msg:
+        - "{{ merged['commands'] | symmetric_difference(result['commands']) | length == 0 }}"
+        - "{{ merged['before'] | symmetric_difference(result['before']) | length == 0 }}"
+        - "{{ merged['after'] | symmetric_difference(result['after']) | length == 0 }}"
+
 
     - name: Assert that correct set of commands were generated
       assert:
         that:
-          - "{{ expected_output | symmetric_difference(result['after']) |length == 0 }}"
+          - "{{ merged['commands'] | symmetric_difference(result['commands']) | length == 0 }}"
+
+    - name: Assert that before dicts are correctly generated
+      assert:
+        that:
+          - "{{ merged['before'] | symmetric_difference(result['before']) | length == 0 }}"
+
+    - name: Assert that after dict is correctly generated
+      assert:
+        that:
+          - "{{ merged['after'] | symmetric_difference(result['after']) | length == 0 }}"
 
     - name: Merge provided configuration with device configuration (IDEMPOTENT)
       ios_interfaces: *merged
